% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UShurricane.R
\docType{data}
\name{UShurricane}
\alias{UShurricane}
\title{US hurricane data}
\format{Data frame with 32060 rows and 3 columns}
\description{
US hurricane data provided by Peter Taylor and Dickie Whitaker.
}
\details{
\itemize{
  \item EventID. ID of 32060 events.
  \item Rate. Annual rate of occurrence.
  \item Loss. Loss associated to each event measured in $.
}
}
\keyword{datasets}

